extract CheckoutWrapper from Checkout, in preparation for future PR
