+++ paywall/src/components/content/CheckoutContent.tsx
@@ -2,7 +2,8 @@ import React, { Fragment } from 'react'
 import Head from 'next/head'
 
 import { pageTitle, ETHEREUM_NETWORKS_NAMES } from '../../constants'
-import Checkout, { CheckoutWrapper } from '../checkout/Checkout'
+import Checkout from '../checkout/Checkout'
+import CheckoutWrapper from '../checkout/CheckoutWrapper'
 import NoWallet from '../checkout/NoWallet'
 import useBlockchainData from '../../hooks/useBlockchainData'
 import useWindow from '../../hooks/browser/useWindow'
